backend/pyproject.toml
[tool.poetry]
name = "CubeOnbackend"
version = "0.1.0"
description = ""
authors = ["artyomklg <artem0801200269@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.97.0"}
sqlalchemy = "^2.0.16"
asyncpg = "^0.27.0"
alembic = "^1.11.1"
async-keycloak = "^0.28.6"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
flake8 = "^6.0.0"
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
